verbose: cast time_t to long long
On Linux and FreeBSD, a time_t can be either a signed 32 or 64-bit value.
OpenBSD sets a time_t to a 64-bit value. When printing the epoch time,
always use a 64-bit value:

~~~
pseudocron.c:180:42: warning: format specifies type 'long' but the argument has type 'time_t' (aka 'long long') [-Wformat]
    (void)fprintf(stderr, "now[%ld]=%s", now, ctime(&now));
                               ~~~       ^~~
                               %lld
pseudocron.c:181:43: warning: format specifies type 'long' but the argument has type 'time_t' (aka 'long long') [-Wformat]
    (void)fprintf(stderr, "next[%ld]=%s", next, ctime(&next));
                                ~~~       ^~~~
                                %lld
pseudocron.c:201:43: warning: format specifies type 'long' but the argument has type 'time_t' (aka 'long long') [-Wformat]
    (void)fprintf(stderr, "exit[%ld]=%s", now, ctime(&now));
                                ~~~       ^~~
                                %lld
3 warnings generated.
~~~

+++ pseudocron.c
@@ -177,8 +177,8 @@ main(int argc, char *argv[])
     err(EXIT_FAILURE, "time");
 
   if (verbose > 0) {
-    (void)fprintf(stderr, "now[%ld]=%s", now, ctime(&now));
-    (void)fprintf(stderr, "next[%ld]=%s", next, ctime(&next));
+    (void)fprintf(stderr, "now[%lld]=%s", (long long)now, ctime(&now));
+    (void)fprintf(stderr, "next[%lld]=%s", (long long)next, ctime(&next));
   }
 
   diff = difftime(next, now);
@@ -198,7 +198,7 @@ main(int argc, char *argv[])
     now = time(NULL);
     if (now == -1)
       err(EXIT_FAILURE, "time");
-    (void)fprintf(stderr, "exit[%ld]=%s", now, ctime(&now));
+    (void)fprintf(stderr, "exit[%lld]=%s", (long long)now, ctime(&now));
   }
 
   return 0;
